# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Register an account",
                "operationId": "ca7c0cb6772de9bffbb8c896980cebc8",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Login to the account",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter your email address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Please enter your password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/get-post-user": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a post resource",
                "operationId": "8416676b82b847fc030bbe276c595d13",
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/show/post/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a post resource",
                "operationId": "73546f3520daadc3e8b494a911b52ccd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/loggout": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Logout account in facebook",
                "operationId": "f0acd86b52674e8e746118f66f22be27",
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/view/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "view profile",
                "operationId": "540de19e09f0c54d20ac8fda7e7c5daa",
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updatPl": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Confirm your friend",
                "operationId": "f10ec4a60c8fef5dda5977c30bb6952f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    }
}